+++ lib/selecto_mix/introspector/ecto.ex
@@ -155,7 +155,7 @@ defmodule SelectoMix.Introspector.Ecto do
       :time -> :time
       :utc_datetime -> :utc_datetime
       :naive_datetime -> :naive_datetime
-      :map -> :jsonb
+      :map -> :json
       {:array, inner_type} -> {:array, map_ecto_type_to_selecto(inner_type)}
       {Ecto.Enum, _values} -> :string
       {:parameterized, Ecto.Enum, _values} -> :string